Lien modèle vue

+++ src/com/pi4/duet/model/Point.java
@@ -1,18 +1,19 @@
 package com.pi4.duet.model;
 
+
 public class Point {
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() { return x; }
-	public int getY() { return y; }
+	public double getX() { return x; }
+	public double getY() { return y; }
 	
-	public void setX(int x) { this.x = x; }
-	public void setY(int y ) { this.y = y; }
+	public void setX(double x) { this.x = x; }
+	public void setY(double y ) { this.y = y; }
 }
